CA-146140:interface-rename fails if eth namelen=16
net-rename-sideways.sh and ifrename/util.py create temporary names for NIC's in
trying to rename them. When the temporary name's length is >= 16, ip link rename
command fails. This will leave the eth devices with names not compliant to the
udev and interface-rename rules.

This change restricts <num> in side-<num>-<ethXX> to 4 bytes. This will keep the
eth devices name length to < 16 for up to eth99.

If in future eth100 or bigger name is a possible, length of <num> should be
further reduced.