CA-238308: Order new multi-NIC devices by MAC
Where PCI functions have multiple NICs, biosdevname will return an undefined order
as it uses the PCI bus location as the key for both the SMBIOS label and the sysfs
label.  Relying on biosdevname in this scenario can mismatch with expectations -
which are typically that the devices are in MAC order.

Therefore keep the ordering according to the MAC specifically in the case of
multi-NIC devices.