CA-246490: version: Change the build number to a build identifier
The build number is now a build identifier, so change the code to work
with this. Drop the build suffix handling.  The build identifier is not
useful for version comparisons, so don't use it in comparisons.

This results in a slight change in behavior in the host installer.
Previously, it would not be possible to upgrade from a later build to an
earlier build of the same platform version. After this change, it would
be possible. This shouldn't be an issue as within a platform version are
only used for testing.