CP-30501: Add API to get dom0 default memory by version
XC needs to know how much memory dom0 will get after an RPU but an RPU
might not always be to the latest version. Therefore, restore the
previous version of the calculation under a new name and add a new API
to get the default dom0 memory for a particular platform version.