CA-339540: Fail NFS mounts faster
If the user tries to RPU from an NFS server that doesn't exist or is not
responding, the mount takes several minutes to timeout which results in
a poor user experience in XenCenter. Set some NFS options so that the
mount fails after 30 seconds.

The options are:

timeo=100 - Wait 10 seconds before retrying an NFS request
retrans=1 - Retry NFS requests once
retry=0 - If the initial mount fails, don't retry