cpiofile: drop CpioFileCompat.write()
This implementation of CpioFileCompat.write cannot work with python
2.7, as adding new attributes to a ZipInfo triggers an exception.
`tarfile.py` in Python 2.7 has a fix, but this method is clearly not
used, so there's no reason to keep it, especially as its use of
StringIO should likely switch to BytesIO and we would need to write
tests for this.

Also note that in python3 the TarFileCompat class is removed.  We could
probably just drop our CpioFileCompat completely, if we're sure noone uses
it, but it likely only exists because there was a TarFileCompat.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>