python3: use six.string_types not version-dependant types
Apply "Use of `unicode` needed to be immediately handled" by Yann.

Reserved "but a few checks relying on `str` could become insufficient
in python2 with the larger usage of unicode strings." for applying
optionally these not-needed later, because they could change behavior
and if needed, they would belong to these other commits. And, after
21 commits from Yann and my work on to of that, that didn't appear.

Co-authored-by: Yann Dirson <yann.dirson@vates.fr>
Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>