xcp.repository: switch from ConfigParser to configparser
This is supposed to be just a module renaming to conform to PEP8, see
https://docs.python.org/3/whatsnew/3.0.html#library-changes

The SafeConfigParser class has been renamed to ConfigParser in Python
3.2, and backported as addon package.  The `readfp` method now
triggers a deprecation warning to replace it with `read_file`.

Originally authored by Yann Dirson,
Updated by Bernhard Kaindl to no longer need the WIP marker,
to unblock the Python3 work, with more tests to be added.

Co-authored-by: Yann Dirson <yann.dirson@vates.fr>
Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>