accessor: Suppress pylint warnings for open(): encoding and use-with
Because the upcoming commits enable pylint to be more accessible,
fix the errors which would be shown around suggesting to pass
encoding to open() calls.

The previous patch changed the default for the mode parameter to
"rb"/"wb" in order to fix reading from binary files with Python3.

This is also conistent with HTTPAccessor only being able to read
bytes because urllib cannot decompress HTTP gzip encodings on the
fly, and a conversion from the decompressed data to unicode sting
can only happen after decompression.

Because of the need to default to binary mode, the pylint warning
to pass encoding= to open() does not apply to these calls which
are setup by their calling mode to use binary mode by default.

The next commit will enforce binary mode unless encoding is specified.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>