accessor: enforce binary mode unless encoding is specified
The previous patch supresses the pylint warning that encoding= should
be bassed to open calls because the default of open() is text mode,
which means decoding on reads and encoding on writes.

Until this commit, this means that it would be the responsibilty
of each callers to specify an encoding and setup handling decoding
errors when they want to use file-based accessors in text mode.

To be safe against mis-use like `openAccess(file, mode="r")`
(note: the mode parameter is new), we could ignore that request
and enforce binary mode unless encoding was passed as well.

Summary: Use bytes whenever possible as this avoids any errors around
coding/conversion of bytes to Unicode strings and back to bytes.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>