.pci.PCIIds.read(): Fix reading pci.ids with UTF-8 chars and test it
The ACK XAPI plugin calls xcp.pci.PCIIds.read() without an UTF-8 locale set.

Problem:

Because there are Unicode characters ('²') in `/usr/share/hwdata/pci.ids`,
it needs to be opened with UTF-8 decoding set up for `read()` to succeed.

But, because XAPI plugins are not started with the locale set to an UTF-8
locale this fails on Python 3.6 because in this case, it defaults to using
the `ascii` codec which does not support `ord()` >= 128.

Thus, `xcp.pci.PCIIds.read()` fails to read the PCI IDs database when no
locale is set on Python 3.6.

Fix:

Wrap the open() call using a function which enables the UTF-8 encoding
when run on Python 3.0 or newer.

The use of the wrapper function also fixes the pylint warning
`unspecified-encoding` for not passing `encoding="utf-8"`.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>