net.biosdevname: For Py3, add universal_newlines=True to Popen()
Besides updating the mock test, also adds a new test using unshare()
to switch the test case into a new user namespace (biosdevname needs
to run as root).

Due to a known issue with pytest-forked (a part of pytest-xdist),
the new test needs to be in a dedicated file.

Because of the fixes, biosdevname and cpiofile no longer need to be
excluded from the main pytype check run by run-pytype.py.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>