tox.ini: Add requires for the tox envrionment itself and small updates
I re-installed all my local Python environments and this forced me
to notice how critical the top-level tox.ini:requires= config for
the dependencies of tox itself are.

Without these taken care of, developers would run in to errors which
are not easy to undertand and fix.

The problem is that for working with Python versions which are EOL
since a long time like Python2.7 and 3.6, nobody cares any more and
the maintainers are deploying code which is not compatible with those.

For example, to current versions of virtualenv use Python features
which are only present in Python 3.7 and newer like dataclasses and
from __future__ import annotations which 3.6 does not have and this
then breaks the detection or the setup of environment of these older
Python versions.

`tox.ini:requires=` now takes care of, that even if your enviroment
is not completely on the required versions described in README.md,
tox will still not fail in obscure ways because it will run in it's
own virtualenv which it sets up itself.

The other changes are just to get better logs from CI and get the
interesting logs on first.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>