xcp/pci.py: Update utf8open() to open_with_codec_handling()
In preparation for adding adding of encoding= and errors= to other
users of open() (logger, environ, bootloader, dynfamic, static):

Add the codec to use an optional parameter in case we want to use
another encoding (e.g. ascii) in some of the user users of open()
and rename tf8open() to open_with_codec_handling().

Also enable strict pyright(CLI of Pylance, the checker in VS Code)
checking for xcp/compat.py and add the stubs directory for mypy.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>