xcp/cpiofile.py: Add type comments to fix static type checking
This adds type comments to fix static type checking by mypy & pyright.

The goal is to be able to use the pyright type checker which is
used by the language backend of the Python extension of vscode.

It can be also used standalone for CI/CD, even in GitHub actions.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>