xcp.pci.pci_sbdfi_to_nic: Silence errors from static type checkers
xcp.pci.pci_sbdfi_to_nic's behavior is to raise Exceptions
on any "not found" condition and callers may rely on it.

Fix static type checking errors (mypy, pyright) by checking the
match and raise the expected Exception in that case match is None.
Else continue with match asserted to be valid for type checking.

Because a verified-to-work check of the Exception strings was
added by the previous commit, we can use the line formatting
used by the black formatter to wrap the long Exception line.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>