xcp/accessor.py: openText(): Context manager to read text from addr
Add `Accessor.openText()` as context manager to wrap openAddress() with
`io.TextIOWrapper` (if needed). It yields IO[str] or False, closing the
created TextIOWrapper and the underlying file on leaving the context.

It is intented for use in xcp.repository and any other user requiring
text as string without having to call .decode() on the returned file
object. It also closes the underlying file object on leaving the context.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>