xcp/**/*.py: Remove shebang which was part of the copyright header
As discussed in https://github.com/xenserver/python-libs/pull/58:

Only files which will be execute by invoking them directly require
the shebang. If a library module would be used as a script as well,
the file could have to contain a __name__ guard to prevent it to
be executed on importing it:

if __name__ == "__main__":
    main()

None of the files in xcp/* have such __name__ guard, hence we can
safely cleanup the excess shebangs in xcp. Even the copyright headers
in the otherwise empty __init__.py files have the shebangs, so it looks
like the shebangs were just added as part of the copyright header.

It could have been used as a marker for editors to switch to
Python mode, about stackoverflow agrees that that's not needed.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>