test_bootloader: Fix sporadic failure
This test is broken in multiple ways. Either:

1) Execution of diff races against the named temporary file being
removed when the context manager is left, resulting in no output from
diff and the test "passes" (diff exits with status 2).

2) diff generates output and the call to assertRegexpMatches() fails
   because it was renamed in Python 3.2.

Fix this by ensuring diff finishes before leaving the context manager,
asserting the exact diff expected, and checking the return code.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>